refactor(backend): split route modules from app
