+++ src/routes/mod.rs
@@ -1,0 +1,6 @@
+pub mod channels;
+pub mod health;
+
+// Explicit exports - no wildcards
+pub use channels::{ChannelState, LiveStatusState, channel_routes, live_status_routes};
+pub use health::{APP_VERSION, health_routes};
